refactor: moving amm query functionality to the query folder
