+++ liquidator/src/main.ts
@@ -9,9 +9,9 @@ import { getConfig as getRoverConfig } from './rover/config/osmosis'
 import { RoverExecutor } from './rover/RoverExecutor'
 import { getSecretManager } from './secretManager'
 import { Network } from './types/network'
-import { PoolDataProviderInterface } from './amm/PoolDataProviderInterface.js'
-import { OsmosisPoolProvider } from './amm/OsmosisPoolProvider.js'
-import { AstroportPoolProvider } from './amm/AstroportPoolProvider.js'
+import { PoolDataProviderInterface } from './query/amm/PoolDataProviderInterface.js'
+import { OsmosisPoolProvider } from './query/amm/OsmosisPoolProvider.js'
+import { AstroportPoolProvider } from './query/amm/AstroportPoolProvider.js'
 
 const REDBANK = 'Redbank'
 const ROVER = 'Rover'
