+++ liquidator/src/query/amm/AstroportPoolProvider.ts
@@ -1,5 +1,5 @@
-import { sleep } from "../helpers";
-import { Pool, PoolAsset } from "../types/Pool";
+import { sleep } from "../../helpers";
+import { Pool, PoolAsset } from "../../types/Pool";
 import { PoolDataProviderInterface } from "./PoolDataProviderInterface";
 import fetch from 'cross-fetch'
 import { Asset, AssetInfo, AssetInfoNative, ContractQueryPairs, ContractQueryPool, Pair, PoolResponseData, Query, ResponseData } from "./types/AstroportTypes";
